requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pdf2image==1.16.3
httpx==0.25.2
pydantic==2.5.0